Fix home progress

+++ app/src/main/java/ru/frogogo/whitelabel/data/model/ui/home/HomeProgressUiModel.kt
@@ -1,10 +1,11 @@
 package ru.frogogo.whitelabel.data.model.ui.home
 
 import ru.frogogo.whitelabel.core.recycler.RecyclerViewItem
+import java.math.BigDecimal
 
 data class HomeProgressUiModel(
-  val current: Int,
-  val target: Int,
+  val current: BigDecimal,
+  val target: BigDecimal,
 ) : RecyclerViewItem {
 
   override fun getId(): Any = "item_coupon_progress"
